flight_train_claude/requirements.txt
windows-curses==2.3.1; platform_system == "Windows"
numpy==1.24.3
pyyaml==6.0.1 